[#156322494] Adaptando crud de eventos da API pra aglutinar os eventos

+++ app/api/v1/remove_after_changes.rb
@@ -15,6 +15,8 @@ module V1
               user       = verify_or_create_user(allocation[:cpf])
               profile_id = get_profile_id(allocation[:perfil])
 
+              raise "user #{allocation[:cpf]} doesn't exist" if user.blank? || user.id.blank?
+
               destination = get_destination(allocation[:codDisciplina], allocation[:codGraduacao], allocation[:nomeTurma], (allocation[:periodo].blank? ? allocation[:ano] : "#{allocation[:ano]}.#{allocation[:periodo]}"))
 
               destination.allocate_user(user.id, profile_id)
@@ -88,6 +90,11 @@ module V1
 
             begin
               group = nil
+              if load_group[:name].blank?
+                load_group[:name] = load_group[:codigo]
+                load_group[:code] = load_group[:codigo]
+              end
+
               ActiveRecord::Base.transaction do
                 semester = verify_or_create_semester(semester_name, offer_period)
                 offer    = verify_or_create_offer(semester, {curriculum_unit_id: uc.id, course_id: course.id}, offer_period)
@@ -111,6 +118,7 @@ module V1
               @student_profile  = 1 # Aluno => 1
 
               @groups = @groups.collect do |group_info|
+                group_info["nome"] = group_info["codigo"] if group_info["nome"].blank?
                 get_group_by_names(group_info["codDisciplina"], group_info["codGraduacao"], group_info["nome"], (group_info["periodo"].blank? ? group_info["ano"] : "#{group_info["ano"]}.#{group_info["periodo"]}")) unless group_info["codDisciplina"] == 78
               end # Se cód. graduação for 78, desconsidera (por hora, vem por engano).
 
@@ -136,6 +144,19 @@ module V1
             end
 
           end # segment
+
+          # PUT load/groups/cancel_students_enrollments
+          params{ requires :semester, type: String }
+          put :cancel_students_enrollments do
+            begin
+              ActiveRecord::Base.transaction do
+                semester = Semester.find_by_name(params[:semester])
+                cancel_all_allocations(1, semester.id) # Aluno => 1
+              end
+
+              { ok: :ok }
+            end
+          end
 
           # GET load/groups/enrollments
           params { requires :codDisciplina, :codGraduacao, :nomeTurma, :periodo, :ano, type: String }
@@ -175,13 +196,17 @@ module V1
           end
           put "/:id" do
             begin
-              event = ScheduleEvent.find(params[:id])
+              # event = ScheduleEvent.find(params[:id])
+              ac = AcademicAllocation.where(id: params[:id], academic_tool_type: 'ScheduleEvent').first
+              raise ActiveRecord::RecordNotFound if ac.blank?
+              event = ScheduleEvent.find(ac.academic_tool_id)
 
               ActiveRecord::Base.transaction do
                 start_hour, end_hour = params[:HoraInicio].split(":"), params[:HoraFim].split(":")
-                event.schedule.update_attributes! start_date: params[:Data], end_date: params[:Data]
-                event.api = true
-                event.update_attributes! start_hour: [start_hour[0], start_hour[1]].join(":"), end_hour: [end_hour[0], end_hour[1]].join(":")
+                create_event({event: {date: params[:Data], start: params[:HoraInicio], end: params[:HoraFim], title: event.title, type_event: event.type_event}}, ac.allocation_tag.group.offer.allocation_tag.related, nil, ac, (event.academic_allocations.count == 1 ? event : nil))
+                # event.schedule.update_attributes! start_date: params[:Data], end_date: params[:Data]
+                # event.api = true
+                # event.update_attributes! start_hour: [start_hour[0], start_hour[1]].join(":"), end_hour: [end_hour[0], end_hour[1]].join(":")
               end
 
               { ok: :ok }
@@ -207,10 +232,11 @@ module V1
             begin
               ActiveRecord::Base.transaction do
                 offer = get_offer(params[:CodigoDisciplina], params[:CodigoCurso], params[:Periodo])
-                params[:Turmas].each do |group_name|
-                  group = get_offer_group(offer, group_name)
-                  group_events << create_event1(get_offer_group(offer, group_name), params[:DataInserida])
-                end
+                group_events = create_event(params, offer.allocation_tag.related, offer)
+                # params[:Turmas].each do |group_name|
+                #   group = get_offer_group(offer, group_name)
+                #   group_events << create_event1(get_offer_group(offer, group_name), params[:DataInserida])
+                # end
               end
 
               group_events
@@ -218,16 +244,28 @@ module V1
 
           end # /
 
-          desc "Remoção de um ou mais eventos"
+          desc "Remoção de um ou mais acs de eventos"
           params { requires :ids, type: String, desc: "Events IDs." }
           delete "/:ids" do
             begin
-              ScheduleEvent.transaction do
-                ScheduleEvent.where(id: params[:ids].split(",")).each do |event|
+              AcademicAllocation.transaction do
+                acs = AcademicAllocation.where(id: params[:ids].split(','), academic_tool_type: 'ScheduleEvent')
+                raise ActiveRecord::RecordNotFound if acs.empty?
+                event = ScheduleEvent.find(acs.first.academic_tool_id)
+
+                if acs.count == event.academic_allocations.count
                   event.api = true
                   raise event.errors.full_messages unless event.destroy
+                else
+                  acs.destroy_all
                 end
               end
+              # ScheduleEvent.transaction do
+              #   ScheduleEvent.where(id: params[:ids].split(",")).each do |event|
+              #     event.api = true
+              #     raise event.errors.full_messages unless event.destroy
+              #   end
+              # end
 
               {ok: :ok}
             end
